Fixed little bugs

+++ controllers/userProfileController.js
@@ -8,9 +8,9 @@ const prisma = new PrismaClient();
 const uploadDp = async (req,res) =>
 {
   const dp = req.file;
-  const { id } = req.params;
+  const { id } = res.user;
   try {
-    if ( !id ) res.status( 400 ).json( { message: "User Id is required" } );
+    if ( !id )return res.status( 400 ).json( { message: "User Id is required" } );
     const user = await prisma.user.findUniqueOrThrow( {
       where:{id}, select:safeUser
     } )
@@ -43,7 +43,7 @@ const uploadDp = async (req,res) =>
 
 const deleteDp = async ( req, res ) =>
 {
-  const { id } = req.params;
+  const { id } = res.user;
   try {
     const user = await prisma.user.findUniqueOrThrow( {
       where:{id}, select:safeUser
@@ -74,15 +74,16 @@ const editProfile = async ( req, res ) =>
 {
   try {
     const { name, phone_number, username,address } = req.body;
-    const { id } = req.params;
-    if ( !id ) res.status( 400 ).json( { message: "User Id is required" } );
+    const { id } = res.user;
+    if ( !id ) return res.status( 400 ).json( { message: "User Id is required" } );
 
     const foundUser = await prisma.user.findUniqueOrThrow( { where: { id } } );
+
 
     foundUser.name = name ? name : foundUser.name;
     foundUser.phone_number = phone_number ? phone_number : foundUser.phone_number;
     foundUser.username = username ? username : foundUser.username;
-    foundUser.address = address ? [ address, ...foundUser.address ] : foundUser.address;
+    foundUser.address = address ? [ {address, id: foundUser.address.length ++}, ...foundUser.address ] : foundUser.address;
     
     const updated = await prisma.user.update( {
       where: { id },
@@ -108,6 +109,7 @@ const deleteAccount = async (req,res) =>
 {
   const { id } = res.user
   try {
+    if ( !id ) return res.status( 400 ).json( { message: "User Id is required" } );
     const user = await prisma.user.findUniqueOrThrow( { where: { id } } );
     await fs.unlink( `./public/images/user/${ user.dp }` );
     await prisma.user.delete( { where: { id } } );
@@ -128,7 +130,7 @@ const updatePassword = async ( req, res ) =>
   const { id } = res.user
   const {newPassword, oldPassword} = req.body
   try {
-    if(!newPassword || !oldPassword) return res.status(400).json({message:"All filed must be entered"})
+    if ( !newPassword || !oldPassword ||!id) return res.status(400).json({message:"All filed must be entered"})
     const user = await prisma.user.findUniqueOrThrow( { where: { id } } );
     const verify = await argon.verify( user.password, oldPassword )
     if ( !verify ) return res.status( 401 ).json( { message: 'Current password is incorrect' } )
@@ -149,5 +151,124 @@ const updatePassword = async ( req, res ) =>
   }
 }
 
+const getAllAddress = async ( req, res ) =>
+{
+  const { id } = res.user
+  try {
+    if ( !id ) return res.status( 400 ).json( { message: "User Id is required" } );
+    const user = await prisma.user.findUniqueOrThrow( { where: { id } } );
+    return res.status(200).json({address:user.address})
+  } catch (e) {
+    if (e instanceof Prisma.PrismaClientKnownRequestError) {
+      if (e.code === "P2025")
+        return res.status(404).json({ message: "User not found" });
+    }
+    return res
+      .status(500)
+      .json({ message: "internal server error", error: e.message });
+  }
+}
 
-module.exports ={uploadDp,editProfile,deleteDp,deleteAccount,updatePassword}
+const editAddress = async ( req, res ) =>
+{
+  const { id } = res.user;
+  const { address, index } = req.body;
+  try {
+    if ( !id ) return res.status( 400 ).json( { message: "User Id is required" } );
+    const user = await prisma.user.findUniqueOrThrow( { where: { id } } );
+    const oldAddress = user.address.find( item => item.id === index )
+    if ( !oldAddress ) return res.status( 400 ).json( { message: "Address was not found" } )
+    oldAddress.address = address;
+    user.address.filter( address => address.id !== index );
+    user.address.push( oldAddress );
+    await prisma.user.update( { where: id, data: user } );
+    return res.status( 202 ).json( { message: "Address was updated successfully", address } );
+  } catch (e) {
+    if (e instanceof Prisma.PrismaClientKnownRequestError) {
+      if (e.code === "P2025")
+        return res.status(404).json({ message: "User not found" });
+    }
+    return res
+      .status(500)
+      .json({ message: "internal server error", error: e.message });
+  }
+}
+
+const deleteAddress = async ( req, res ) =>
+{
+  const { id } = res.user;
+  const { index } = req.params
+  try {
+    if ( !id )return res.status( 400 ).json( { message: "User Id is required" } );
+    if ( !index ) return res.status( 400 ).json( { message: "Address id is required"} );
+    const user = await prisma.user.findFirstOrThrow( { where: { id } } )
+    const oldAddress = user.address.find( item => item.id === index )
+    if ( !oldAddress ) return res.status( 400 ).json( { message: "Address not found" } );
+    user.address.filter( address => address.id !== index );
+    user.address.forEach( address =>
+    {
+      address.id = user.address.indexOf(address) + 1
+    } )
+    await prisma.user.update( { where: { id }, data: user } )
+    return res.status(202).json({message:"Address Deleted", address: user.address})
+    
+  } catch (e) {
+    if (e instanceof Prisma.PrismaClientKnownRequestError) {
+      if (e.code === "P2025")
+        return res.status(404).json({ message: "User not found" });
+    }
+    return res
+      .status(500)
+      .json({ message: "internal server error", error: e.message });
+  }
+}
+
+const addAddress = async ( req, res ) =>
+{
+  const { id } = res.user;
+  try {
+    if ( !id ) return res.status( 400 ).json( { message: "User Id is required" } );
+    const { address } = req.body;
+    if ( !address ) return res.status( 400 ).json( { message: "address is required" } );
+    
+    const user = await prisma.user.findUniqueOrThrow( { where: id } );
+    user.address.push( { address, id: user.address.length++ } );
+    await prisma.user.update( {
+      where: { id },
+      data:user
+    } )
+    return res.status( 202 ).json( { message: "Address added successfully" } );
+  } catch (e) {
+    if (e instanceof Prisma.PrismaClientKnownRequestError) {
+      if (e.code === "P2025")
+        return res.status(404).json({ message: "User not found" });
+    }
+    return res
+      .status(500)
+      .json({ message: "internal server error", error: e.message });
+  }
+}
+
+const deleteAllAddress = async ( req, res ) =>
+{
+  const { id } = res.user;
+  try {
+    if ( !id )return res.status( 400 ).json( { message: "User Id is required" } );
+    const user = await prisma.user.findFirstOrThrow( { where: { id } } )
+    user.address = []
+    await prisma.user.update( { where: { id }, data: user } )
+    return res.status(202).json({message:"Addresses Deleted"})
+    
+  } catch (e) {
+    if (e instanceof Prisma.PrismaClientKnownRequestError) {
+      if (e.code === "P2025")
+        return res.status(404).json({ message: "User not found" });
+    }
+    return res
+      .status(500)
+      .json({ message: "internal server error", error: e.message });
+  }
+}
+
+
+module.exports ={uploadDp,editProfile,deleteDp,deleteAccount,updatePassword,addAddress,editAddress,getAllAddress,deleteAddress,deleteAllAddress}
